requirements.txt
black==24.10.0
click==8.1.7
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
djangorestframework